index better, asset 50 made

+++ programming/resethtmlfolder.py
@@ -12,16 +12,17 @@ def resethtmlfolder(htmlfolderpath):
     
     delcount=0
     for folder in folders:
-        fullpath=f'{htmlfolderpath}/{folder}'
-        
-        try:
-            # shutil.rmtree deletes folder and all contents inside it
-            shutil.rmtree(fullpath)
-            delcount += 1
-        except PermissionError:
-            print(f"Permission denied: {fullpath}\n")
-        except Exception as e:
-            print(f"Failed to delete {fullpath}: {e}")
+        if '50' not in folder and os.path.isdir(folder):
+            fullpath=f'{htmlfolderpath}/{folder}'
+            
+            try:
+                # shutil.rmtree deletes folder and all contents inside it
+                shutil.rmtree(fullpath)
+                delcount += 1
+            except PermissionError:
+                print(f"Permission denied: {fullpath}\n")
+            except Exception as e:
+                print(f"Failed to delete {fullpath}: {e}")
 
 
     print(f'Ran resethtmlfolder.\n{delcount} folders were deleted.')
@@ -29,7 +30,8 @@ def resethtmlfolder(htmlfolderpath):
 def main():
     resethtmlfolder(htmlfolderpath=htmlfolderpath)
 
-active =True
+# this will delete web asset 50 though so be careful
+active =False
 if active:
     if __name__=='__main__':
         main()
